Update App for iOS6 and Deprecated methods
There are a few deprecated methods that I decided to keep for various reasons:
1) minimumFontSize is kept to support older phones
2) Dismissing or presenting view controllers is deprecated, but keeping it for
support of older apps.
3) Largest change is viewDidUnload. viewDidUnload is only called when the
conroller's view is released from memory. Usually called when a low-memory
condition occurs and this view is not needed. In iOS6, viewDidUnload is NEVER
called because views are no longer purged under low-memory conditions. Instead
Apple want's developers to use didReceiveMemoryWarning. To make the app
compatible with older iOS versions, I kept the viewDidUnload method and instead
move the variables that should be set to nil to the didReceiveMemoryWarning to
make iOS 6 happy.