Remove keyboard when user taps UITableView
While a user was inputting their username/password, there was no way to cancel
input and dismiss the keyboard unless the user filled out the information and
sent a request to our server. This is pretty annoying to the user. This commit
allows the user to tap outside of the UITableView cells to dismiss the keyboard