Fix iOS App from crashing when a guidestep has no image
When there was an empty image on a guide step, it would crash the app because
it was looking for the key "type" in the dictionary "media". All guide steps
will have a "media" key, but not a "type" key. Apparently this is changing in
API 1.1 but for now this will do.

What happened before: Tried to access key "type" within the media dictionary
and app would crash.

What happens now: If the key type doesn't exist, lets default to "image" type.
We now also add a background image to the Main Image which is the Ruler image.
If images are found they will be placed as a foreground image. Otherwise
nothing will happen and the Ruler image will display.