Extend private site logic for custom iOS Apps
The logic was there to account for private sites, it just needed some extra
logic to get it working with Custom iOS Apps.

I also modified the easyAppOven script to use spaces instead of tabs (Git hub wasn't
happy with tab characters). The script will now also ask the user running the
script if the site is private or not, and add the correct logic for custom
private iOS apps.