Only enable scrolling on WebViews if we need to
Previously on every guide step and guide intro, we made the UIWebView
scrollable (this is the content that displays the bullets and text about each
step). This was annoying to users swiping through guide steps and getting
caught on text that would scroll. This now only enables the scrolling if we
need to scroll.