Restructure category data/logic for CategoryViewController
On iFixit, we want to keep the distinction between categories and devices.
On Dozuki we want to merge that distinction, and just have categories. For some
context, this code modifies the UITableViewController that is responsible for
presenting a list of categories to the user to drill down and finally select a
guide.I tried to simplify things, I feel they were overly complicated before.
This also makes it easy to make future changes regarding differences between
iFixit/Dozuki, in case Taylor changes his mind =).