Fix Favorites controller animation/behavior
Before on iPad when a user selected the favorites button, it would appear
across the whole screen, instead it only appears on the first view of the split
view controller.

For iPad I also added a curl up animation to mimic a modal presentation, when
really I am pushing the view on and off the navigation stack.

on iPhone we just use a simple modal presentation so I don't have to recreate
the animations.