Fix GuideView Loading/unloading View controllers
We have logic where we destroy view controllers we are no longer using
or we need to create view controllers. This logic was originally in a delegate
method called "scrollViewDidScroll". Problem is this delegate method can be
called upwards of 20+ times per step, so maybe that logic shouldn't have been thrown in
this delegate method. Instead, I moved it to the correct delegate method that
gets called only once.

TLDR; Method that does expensive things used to be called 20+ times per step,
when it should have only been called once. This fixes that.