Fix visual bug when user has large list of bookmarks
When a user has a large list of bookmarks and they logout, it was creating a
very weird visual error. What I found out was happening was even though we
logout, the viewcontroller still has a uitableview to deal with. So when we
logout, I set our bookmarks to be nil and reload the tableview, which will now
correctly release the old data.