Add ability to change title to SVWebViewControllers
We are using a framework for displaying a webview on the tablet, by default it
would rip the document title from the html using javascript. Now we try to use
a title if we have one, otherwise we default to original behavior, this allowed
me to fix the issue where the title was being truncated for it's length.