Make search smarter
Before we only limited search to devices, but now we can search for any
category and present the correct data to the user. The idea is that the user
searches for a category through our API, and once they find a category we drop
them into the correct hierarchy, mimicking as if they navigated their
themselves.