Fix adding guides to category list for slow connections
Problem:
People with slow connections would often be able to navigate through multiple
views, and fire off a good amount of async requests to our api.  By the time
the responses got back, the user be on a completely different view. This would
cause the request to dump it's results in whatever view the current user is
looking at, even if the request did not belong to the view. This foobars
the UITableView's datasource, which is what caused the crash because we
would try to update the tableview to represent the datasource but the
datasource gets corrupted. The Datasource is just the data representation of
the items in the UITableView.

Having a slow connection uncovered this bug. Now I made things more robust to
map each request to it's respective view controller.