Smart-throttle search
We were sending a request to our API after every keypress. This is detremental
for users with a slow connection, and also we do a lot of unnecessary requests
  to our API.

New Behavior:
If the search term is <= 3 characters, then we don't throttle requests.
Once the search term is > 3 characters, we then wait until the user has stopped
typing for 500ms before firing off an api request.