Hardcode the height of category cells
Before, the heightForCellAtIndexPath delegate was not implemented.
In iOS 8, this was causing the heights of freshly created cells to be
like, super high and big and ugly.

So, this sets the height of all the cells in the table view to
43.5, which was the height of a correct cell. It doesn't appear that
any other cells in the controller use a different height.