Fix clipping of the GuideViewController
The view width being reported was always for landscape mode because we
were calling for it before the (rotated) view had been drawn.
This meant the first view of a guide in landscape was clipped to the
width of a portrait view until it was reflowed with a rotation.

To fix this, we just hardcode the view frame if we're on ios 8.
Let's hope it doesn't break ios7.